% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grid_jitter.R
\name{grid_jitter}
\alias{grid_jitter}
\title{Remove over-plotting by gridding point data intelligently}
\usage{
grid_jitter(x, y = NULL, nx = 50, ny = NULL, tol = 5,
  plotresults = TRUE, file = NULL, w = 10, h = 10)
}
\arguments{
\item{x}{Numeric vector or 2 column matrix or data.frame of data points to plot}

\item{y}{Numeric vector, y coordinates matching x (if a vector)}

\item{tol}{Numeric, the maximum distance overplotted points are allowed to move to the nearest vacant grid cell}

\item{plotresults}{Logical, output plots to illustrate point displacements and cell-reallocations}

\item{file}{String, if not NULL the result of plotresults will save to filename instead of rendering in R}

\item{nx/ny}{Numeric, values for the grid dimensions}

\item{w/h}{Numeric, dimensions (inches) of plotresults if output to file (PDF)}
}
\value{
A 2 column matrix of grid-jittered point coordinates
}
\description{
Data jittering reduces overplotting by adding small variances to values. gridJitter removes it entirely by fitting points to a custom grid and then reallocating duplicate points to the nearest vacant cell within a maximum tolerance threshold.  If no vacant cells are available the function will abort.
}
\examples{
# 1. normal distribution
d1 = data.frame(x = rnorm(500), y = rnorm(500))
d1g = grid_jitter(d1, nx=75, ny=75, tol=5)

# 2. faithful dataset
d2 = faithful
d2g = grid_jitter(d2, nx=50, ny=50, tol=3)

# 3. mpg example
d3 = ggplot2::mpg[,c(3,8)]
d3g = grid_jitter(d3, nx=100, ny=100, tol=3)

# 4. diamonds - carat vs price
d4 = ggplot2::diamonds[sample(1:53940, 500),c('carat','price')]
d4g = grid_jitter(d4, nx=300, ny=300, tol=2) # fails
d4g = grid_jitter(d4, nx=300, ny=300, tol=5, plotresults=F)

# 5. US States
d5 = data.frame(x=state.center$x, y=state.center$y, id=state.abb)
d5j = grid_jitter(d5[,1:2], nx=10, ny=8, tol=3, plotresults=T)
cols = sample(colorRampPalette(c('darkblue','blue','lightblue'))(50))
par(mai=rep(.6,4), mfrow=c(1,2))
plot(d5[,1:2], pch=15, cex=2.5, col=cols, asp=2.5, xlab=NA, ylab=NA)
text(d5[,1:2], labels=state.abb, col='white', cex=.7)
plot(d5j, pch=15, cex=2.5, col=cols, asp=2.5, xlab=NA, ylab=NA)
text(d5j, labels=d5$id, col='white', cex=.7)

# 6. to illuste cell reallocation 
d5 = data.frame(x = c(1,rep(50,500), 100), y = c(1,rep(50,500), 100))
d5g = grid_jitter(d5, nx=30, ny=30, tol=15)
}

